+++ src/audiobuffer/stream_decoder.h
@@ -76,7 +76,7 @@ struct FlacInfo {
 	uint32_t sample_rate;
 	uint32_t channels;
 	uint32_t bits_per_sample;
-	uint64_t total_samples;
+	uint32_t total_samples;
 };
 
 struct OggMetadata {
